+++ forms/src/forms.ts
@@ -11,3 +11,4 @@ export * from './misc/utils';
 export * from './misc/types';
 export * from './modules/numberInput.module';
 export * from './modules/hasError.module';
+export * from './modules/requiredClass.module';