+++ forms/src/modules/hasError.module.ts
@@ -2,7 +2,6 @@ import {NgModule} from '@angular/core';
 
 import {HasErrorDirective} from '../directives/hasError/hasError.directive';
 import {GroupHasErrorDirective} from '../directives/groupHasError/groupHasError.directive';
-import {RequiredClassDirective} from '../directives/requiredClass/requiredClass.directive';
 
 /**
  * Module for input validation directives
@@ -10,11 +9,9 @@ import {RequiredClassDirective} from '../directives/requiredClass/requiredClass.
 @NgModule(
 {
     declarations: [HasErrorDirective,
-                   GroupHasErrorDirective,
-                   RequiredClassDirective],
+                   GroupHasErrorDirective],
     exports: [HasErrorDirective,
-              GroupHasErrorDirective,
-              RequiredClassDirective]
+              GroupHasErrorDirective]
 })
 export class HasErrorModule
 {