+++ forms/src/modules/requiredClass.module.ts
@@ -1,0 +1,15 @@
+import {NgModule} from '@angular/core';
+
+import {RequiredClassDirective} from '../directives/requiredClass/requiredClass.directive';
+
+/**
+ * Module for required class directive
+ */
+@NgModule(
+{
+    declarations: [RequiredClassDirective],
+    exports: [RequiredClassDirective]
+})
+export class RequiredClassModule
+{
+}